Make TableSection into its own component
