+++ notes/TODO.md
@@ -1,27 +1,52 @@
-## TODO
+# TODO
 
 ## General
 
-- [ ] Refresh on why and where I'm using the Context API and how it works. Is it needed in this case.
 - [ ] Add iPhone notes idioms to this readme right above this todo.
-- [ ] Comment out the idiom scraper and document it. Like what it does and how to use it. Add it to the repo.
-- [x] go through and clean up the READMEs.
-- [ ] Finish the pern tutorial.
-- [x] Get a copy of the original idioms_nonum.csv file to put in the repo
 
-### Testing
+## Styled components Refactor
+
+- [ ] Merge styled-components into main.
+
+## CSS organization
+
+- [ ] Use color variables
+- [ ] Use sizing variables
+
+## Update, Add functionality
+
+- [ ] Create temp popups for successfull / unsuccessfull add and update idioms
+
+## Users/Login
+
+- [ ] Create login and auth system with auth0.
+
+## Delete Idiom functionality
+
+- [ ] Make sure it works
+- [ ] ask user if they are sure first
+- [ ] Go back to main page on success
+- [ ] Create temp popups for successfull / unsuccessfull delete idioms
+- [ ] Saving user preferences for column visibility.
+
+## Learning
 
 - [ ] Figure out why and how to use Postman
-- [x] Test all the api routes with new wide data.
+- [ ] Finish the pern tutorial.
+- [ ] Refresh on why and where I'm using the Context API and how it works. Is it needed in this case.
 
-### Server
+## Remove old timestamp column from databases
 
-### Client
+- [ ] Remove the old timestamp column on pg admin.
+- [ ] Export the .sql as idioms_test_with_tz_backup.sql.
+- [ ] Burn supabase tables and load the new one.
 
-- [x] Add pagenation feature
-- [x] fix updateIdiom.jsx !!!
+## Convert project to typescript
 
-## Log
+- [ ] Watch youtube about it
+- [ ] Convert project to typescript
+
+# Log
 
 3-21-24:
 
@@ -31,6 +56,7 @@
 - [x] Deleted `idioms_wide`. This table was the original complete table with dates and names.
 - [x] Created `idioms`, `idioms_test`, `idioms_backup`. These tables are all copies of `idioms_wide`.
 - [x] Update database tables.
+- [x] Test all the api routes with new wide data.
 
 4-4-24
 
@@ -57,12 +83,19 @@
 - [ ] Fix If changing the sorting direction of one of the table columns. Go back to pagination page 1.
 - [ ] Fix Column show dropdown is behind selected pagination page link when table is very short .
 
-9-12-24
+- [x] Add pagenation feature
+- [x] fix updateIdiom.jsx !!!
 
-- [ ] Fix delete idiom.
+9-12-24
 
 9-21-24
 
 - [x] Add 7 hours to the timestamp column in all rows of idioms_test
 - [x] Update the addIdiom form to adjust to this change
 - [x] Update the updateIdiom form to adjust to this change
+
+9-23-24
+
+# Doubles
+
+- 503 Dont cry over spilt milk
